+++ resources/views/themes/builder-toolbar.blade.php
@@ -193,21 +193,40 @@
         @if(auth('web')->user())
             <x-splade-data :default="['showOptions' => false]">
                 <div @mouseover="data.showOptions = true" @mouseleave="data.showOptions = false" class="relative">
-                    @include($section['section'], ['page' => $page, 'section' => $section, 'sectionID' =>$section['uuid']])
-                    <div class="my-4 mx-4 absolute z-0 bottom-0 left-0" v-if="data.showOptions">
-                        @if($section['form'])
-                            <x-tomato-admin-button  href="{{route('admin.pages.meta', $page->id) . '?section='. $section['uuid']}}" modal>
-                                <i class="bx bx-edit"></i>
+                    @if(view()->exists($section['section']))
+                        @include($section['section'], ['page' => $page, 'section' => $section, 'sectionID' =>$section['uuid']])
+                        <div class="my-4 mx-4 absolute z-0 bottom-0 left-0" v-if="data.showOptions">
+                            @if($section['form'])
+                                <x-tomato-admin-button  href="{{route('admin.pages.meta', $page->id) . '?section='. $section['uuid']}}" modal>
+                                    <i class="bx bx-edit"></i>
+                                </x-tomato-admin-button>
+                            @endif
+                            <x-tomato-admin-button danger confirm method="DELETE" :data="['section' => $section['uuid']]" href="{{route('admin.pages.remove', $page->id)}}">
+                                <i class="bx bx-trash"></i>
                             </x-tomato-admin-button>
-                        @endif
-                        <x-tomato-admin-button danger confirm method="DELETE" :data="['section' => $section['uuid']]" href="{{route('admin.pages.remove', $page->id)}}">
-                            <i class="bx bx-trash"></i>
-                        </x-tomato-admin-button>
-                    </div>
+                        </div>
+                    @else
+                        <div class="cursor-move flex flex-col gap-4 m-4 text-danger-500 items-center text-center justifiy-center w-full border rounded-lg p-4">
+                            <x-heroicon-s-x-circle class="w-12 h-12" />
+                            {{__('View Not Exists Please Delete IT!')}}
+                        </div>
+                        <div class="my-4 mx-4 absolute z-0 bottom-0 left-0" v-if="data.showOptions">
+                            <x-tomato-admin-button danger confirm method="DELETE" :data="['section' => $section['uuid']]" href="{{route('admin.pages.remove', $page->id)}}">
+                                <i class="bx bx-trash"></i>
+                            </x-tomato-admin-button>
+                        </div>
+                    @endif
                 </div>
             </x-splade-data>
         @else
-            @include($section['section'], ['page' => $page, 'section' => $section, 'sectionID' =>$section['uuid']])
+            @if(view()->exists($section['section']))
+                @include($section['section'], ['page' => $page, 'section' => $section, 'sectionID' =>$section['uuid']])
+            @else
+                <div class="cursor-move flex flex-col gap-4 m-4 text-danger-500 items-center text-center justifiy-center w-full border rounded-lg p-4">
+                    <x-heroicon-s-x-circle class="w-12 h-12" />
+                    {{__('View Not Exists Please Delete IT!')}}
+                </div>
+            @endif
         @endif
     @endforeach
 @else
